+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentAndroidEventTest.kt
@@ -7,7 +7,7 @@ import com.google.common.truth.Truth.assertThat
 import com.instacart.formula.android.ActivityStore
 import com.instacart.formula.android.Feature
 import com.instacart.formula.android.FeatureFactory
-import com.instacart.formula.android.FragmentFlowStore
+import com.instacart.formula.android.FragmentStore
 import com.instacart.formula.android.ViewFactory
 import com.instacart.formula.android.events.ActivityResult
 import com.instacart.formula.test.TestFragmentActivity
@@ -31,7 +31,7 @@ class FragmentAndroidEventTest {
                         configureActivity = {
                             it.initialContract = TestLifecycleKey()
                         },
-                        fragmentStore = FragmentFlowStore.init {
+                        fragmentStore = FragmentStore.init {
                             val featureFactory = object : FeatureFactory<Unit, TestLifecycleKey> {
                                 override fun initialize(dependencies: Unit, key: TestLifecycleKey): Feature {
                                     return Feature(