+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentLifecycleStateTest.kt
@@ -8,7 +8,7 @@ import com.instacart.formula.android.ActivityStore
 import com.instacart.formula.android.FormulaFragment
 import com.instacart.formula.android.ActivityStoreContext
 import com.instacart.formula.android.FeatureFactory
-import com.instacart.formula.android.FragmentFlowStore
+import com.instacart.formula.android.FragmentStore
 import com.instacart.formula.android.FragmentKey
 import com.instacart.formula.test.TestKey
 import com.instacart.formula.test.TestKeyWithId
@@ -38,7 +38,7 @@ class FragmentLifecycleStateTest {
                         configureActivity = {
                             it.initialContract = TestKey()
                         },
-                        fragmentStore = FragmentFlowStore.init {
+                        fragmentStore = FragmentStore.init {
                             bind(featureFactory<TestKey>(this@activity))
                             bind(featureFactory<TestKeyWithId>(this@activity))
                         }