+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Replace `implementation` function with a value
 - **Breaking**: Removed `FlowFactory` and `Flow`
 - **Breaking**: Replace `FragmentStoreBuilder` with `FeaturesBuilder`
+- **Breaking**: Rename `FragmentFlowStore` to `FragmentStore`, `FragmentState` to `FragmentOutput`, `FragmentFlowState` to `FragmentState`
 
 ## [0.7.1] - June 28, 2022
 - **Breaking**: Rename `FragmentBindingBuilder` to `FragmentStoreBuilder`